+++ basiccheck/slice_test.go
@@ -7,6 +7,30 @@ import (
 	"github.com/jeremmfr/go-utils/basiccheck"
 )
 
+func TestOneOfSliceWith(t *testing.T) {
+	sliceOfString := []string{}
+
+	if basiccheck.OneOfSliceWith(sliceOfString, func(s string) bool {
+		return strings.HasPrefix(s, "b")
+	}) {
+		t.Errorf("OneOfSliceWith found prefix 'b' in empty slice")
+	}
+
+	sliceOfString = append(sliceOfString, []string{"foo", "baz", "bar"}...)
+	if !basiccheck.OneOfSliceWith(sliceOfString, func(s string) bool {
+		return strings.HasPrefix(s, "b")
+	}) {
+		t.Errorf("OneOfSliceWith didn't find prefix 'b' in one of %v", sliceOfString)
+	}
+
+	// find a string without all lowercase letters
+	if basiccheck.OneOfSliceWith(sliceOfString, func(s string) bool {
+		return strings.ToLower(s) != s
+	}) {
+		t.Errorf("OneOfSliceWith found a capital letter in one of %v", sliceOfString)
+	}
+}
+
 func TestStringInSlice(t *testing.T) {
 	sliceOfString := []string{"foo", "bar"}