+++ basiccheck/string.go
@@ -6,11 +6,7 @@ import (
 
 // StringHasOneOfPrefixes check if string has one of prefix list.
 func StringHasOneOfPrefixes(s string, list []string) bool {
-	for _, item := range list {
-		if strings.HasPrefix(s, item) {
-			return true
-		}
-	}
-
-	return false
+	return OneOfSliceWith(list, func(v string) bool {
+		return strings.HasPrefix(s, v)
+	})
 }